Improve the TodoJpaResource

+++ src/main/java/com/in28minutes/rest/webservices/restfulwebservices/todo/TodoJpaResource.java
@@ -1,7 +1,9 @@
 package com.in28minutes.rest.webservices.restfulwebservices.todo;
 
+import java.net.URI;
 import java.util.List;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,6 +12,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import com.in28minutes.rest.webservices.restfulwebservices.todo.repository.TodoRepository;
 
@@ -29,33 +33,53 @@ public class TodoJpaResource {
 	
 	@GetMapping("/users/{username}/todos/{id}")
 	public Todo retrieveTodo(@PathVariable String username, 
-			@PathVariable int id) {
-		return todoRepository.findById(id).get();
+			@PathVariable Integer id) {
+		return todoRepository.findById(id)
+				.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Todo not found"));
 	}
 	
 	@DeleteMapping("/users/{username}/todos/{id}")
 	public ResponseEntity<Void> deleteTodo(@PathVariable String username, 
-			@PathVariable int id) {
+			@PathVariable Integer id) {
+		
+		if(!todoRepository.existsById(id)) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Todo not found");
+		}
+		
 		todoRepository.deleteById(id);
 		
 		return ResponseEntity.noContent().build();
 	}
 	
 	@PutMapping("/users/{username}/todos/{id}")
-	public Todo updateTodo(@PathVariable String username, 
-			@PathVariable int id, @RequestBody Todo todo) {
-		todoRepository.save(todo);
+	public ResponseEntity<Todo> updateTodo(@PathVariable String username, 
+			@PathVariable Integer id, @RequestBody Todo todo) {
 		
-		return todo;
+		if(!todoRepository.existsById(id)) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Todo not found");
+		}
+		
+		todo.setUsername(username);
+		todo.setId(id);
+		Todo updatedTodo = todoRepository.save(todo);
+		
+		return ResponseEntity.ok(updatedTodo);
 	}
 	
 	@PostMapping("/users/{username}/todos")
-	public Todo createTodo(@PathVariable String username, 
+	public ResponseEntity<Todo> createTodo(@PathVariable String username, 
 			@RequestBody Todo todo) {
 		todo.setUsername(username);
 		todo.setId(null);
 		
-		return todoRepository.save(todo);
+		Todo savedTodo = todoRepository.save(todo);
+		URI location = ServletUriComponentsBuilder
+				.fromCurrentRequest()
+				.path("/{id}")
+				.buildAndExpand(savedTodo.getId())
+				.toUri();
+		
+		return ResponseEntity.created(location).body(savedTodo);
 	}
 	
 }
